requirements.txt
numpy
pytorch
scikit-image
torchmetrics
torchvision
